Por fin funciona!!

+++ app/src/main/java/chl/ancud/m6_individual5/data/remote/Terreno.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Terreno(
     val id: String,
-    @SerializedName("price") val precio: Int,
+    @SerializedName("price") val precio: Long,
     @SerializedName("type") val tipo: String,
     @SerializedName("img_src") val imagen: String
 )
